querythrottler: tidy loadshed strategy naming and test placement
- Rename the strategy's gate map field gates -> snakes to match the loadshed
  vocabulary used throughout the package.
- Inline the single-use no-op admission release in AcquireAdmission.
- Move TestLiveCapacity into tabletserver_test.go and drop the standalone
  capacity_test.go, avoiding a near-empty file's license/import boilerplate.
- Drop a redundant field comment on Deps.PoolCapacities.

AI disclosure: Claude Code assisted with development. Every line of code was either written by or carefully reviewed by me :)
